// src/shared/enum/horoscope.enum.ts
export enum Horoscope {
  AQUARIUS = 'aquarius',
  PISCES = 'pisces',
  ARIES = 'aries',
  TAURUS = 'taurus',
  GEMINI = 'gemini',
  CANCER = 'cancer',
  LEO = 'leo',
  VIRGO = 'virgo',
  LIBRA = 'libra',
  SCORPIUS = 'scorpius',
  SAGITTARIUS = 'sagittarius',
  CAPRICORNUS = 'capricornus',
}